use dart runtime bundled with sidekick
